Should dump line numbers as binary

+++ sources/linenumber.hpp
@@ -2,6 +2,6 @@
 #define LINENUMBER_HPP
 
 // The line number nested type name.
-typedef int64_t LineNumber;
+typedef uint32_t LineNumber;
 
 #endif
